extracted Garage to its own file, SingleLot tries to show AcceptForm for single lot
